fastapi/requirements.txt
tensorflow==2.16.1
fastapi==0.110.0
uvicorn==0.29.0
SQLAlchemy==2.0.29
PyMySQL==1.1.0
scikit-learn==1.3.2
mysql==0.0.3
mysqlclient==2.2.4
APScheduler==3.10.4
pyclustering==0.10.1.2
DateTime==5.2
mangum==0.17.0
PyKakao==0.0.7
python-jose==3.3.0
python-multipart==0.0.9
passlib==1.7.4
bcrypt==4.1.2
haversine==2.8.1
pyfcm==1.5.4
folium==0.16.0
